# openapi.yaml
openapi: 3.0.3
info:
  title: zenproxy API
  description: >
    OpenAI/Anthropic-compatible proxy for OpenCode Zen models.
    Serves /v1/chat/completions (OpenAI), /v1/responses (Responses API),
    /v1/messages (Anthropic Messages API), /v1/models, and a web dashboard.
  version: 0.4.0
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:8020
    description: Local zenproxy instance (default port 8020)
...
paths:
...
  /dashboard:
    get:
      tags: [Dashboard]
      summary: Web dashboard (HTML)
      security:
        - dashboardToken: []
      responses:
        "200":
          description: Dashboard HTML page
          content:
            text/html:
              schema: { type: string }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /dashboard/data:
    get:
      tags: [Dashboard]
      summary: Dashboard metrics JSON
      security:
        - dashboardToken: []
      responses:
        "200":
          description: Live metrics (rate windows, usage, keys, models)
          content:
            application/json:
              schema:
                type: object
                properties:
                  generated_at: { type: string }
                  rate_limit:
                    type: object
                    properties:
                      enabled: { type: boolean }
                      requests_per_minute: { type: integer }
                      tokens_per_minute: { type: integer }
                  rate_windows:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        requests: { type: integer }
                        tokens: { type: integer }
                        requests_per_minute: { type: integer }
                        tokens_per_minute: { type: integer }
                  usage:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        requests: { type: integer }
                        prompt_tokens: { type: integer }
                        completion_tokens: { type: integer }
                        total_tokens: { type: integer }
                        cost_usd: { type: number }
                        last_used: { type: string }
                  keys:
                    type: array
                    items:
                      type: object
                      properties:
                        name: { type: string }
                        hash: { type: string }
                        revoked: { type: boolean }
                        budget_usd: { type: number }
                        spend_usd: { type: number }
                        model_allow:
                          type: array
                          items: { type: string }
                  models:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        is_free: { type: boolean }
                        is_go: { type: boolean }
                        is_alias: { type: boolean }
                        upstream: { type: string }
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Managed key (zp_...) or OpenCode key (sk-... with optional zen:/go:
        prefix). Managed keys enforce per-key rate limits, model allowlists,
        and budget caps.
    xApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Alternative header for the same key.
    publicTier:
      type: apiKey
      in: header
      name: Authorization
      description: >
        Value "public" (or omitted) when key_auth.allow_public=true — uses the
        free Zen catalog.
    dashboardToken:
      type: apiKey
      in: query
      name: token
      description: Admin token for the dashboard (dashboard_auth.token).
  responses:
    Unauthorized:
      description: Invalid or missing API key
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message: { type: string }
                  type: { type: string, example: authentication_error }
...